Integrate front-end in backend

+++ models/file.go
@@ -3,7 +3,7 @@ package models
 type File struct {
 	Id string `json:"id" bson:"_id"`
 	Name string `json:"name" bson:"name"`
-	Size int `json:"size" bson:"size"`
+	Size int64 `json:"size" bson:"size"`
 }
